refactor: split guardrails remote backend module
